Split up AnmDecoder into engine specific files
